Mesh 371/map account to did (#127)
* Minimal example

* Add some more comments

* More comments

* Enable `UpdateDid`

* SignedExtension at Runtime Level

* KYC mock up

* Fix build

* Fix release build

* UpdateDid move out from lib.rs

* Tools for testing storage

* Extract runtime

* Compilation fix

* CodeReview: Clean `current_did` once transaction is finished

* CodeReview: forwarded_call checks authorization between DIDs

* Add comment about nested calls

* New add/remove signing identities

* Remove some warnings at tests

* Testing `is_authorized_identity`
