+++ primitives/src/did_record.rs
@@ -1,7 +1,7 @@
 use parity_scale_codec::{Decode, Encode};
 use rstd::prelude::Vec;
 
-use crate::{IdentityRole, Key, SigningKey};
+use crate::{IdentityId, IdentityRole, Key, SigningKey};
 
 /// Identity information.
 #[allow(missing_docs)]
@@ -10,6 +10,7 @@ pub struct DidRecord {
     pub roles: Vec<IdentityRole>,
     pub master_key: Key,
     pub signing_keys: Vec<SigningKey>,
+    pub signing_identities: Vec<IdentityId>,
 }
 
 impl DidRecord {
