+++ runtime/src/general_tm.rs
@@ -288,7 +288,11 @@ mod tests {
         AnySignature, Perbill,
     };
     use srml_support::traits::Currency;
-    use srml_support::{assert_ok, impl_outer_origin, parameter_types};
+    use srml_support::{
+        assert_ok,
+        dispatch::{DispatchError, DispatchResult},
+        impl_outer_origin, parameter_types,
+    };
     use std::result::Result;
     use substrate_primitives::{Blake2Hasher, H256};
     use test_client::{self, AccountKeyring};
@@ -426,8 +430,24 @@ mod tests {
         type FullIdentificationOf = ();
     }
 
+    #[derive(codec::Encode, codec::Decode, Debug, Clone, Eq, PartialEq)]
+    pub struct IdentityProposal {
+        pub dummy: u8,
+    }
+
+    impl sr_primitives::traits::Dispatchable for IdentityProposal {
+        type Origin = Origin;
+        type Trait = Test;
+        type Error = DispatchError;
+
+        fn dispatch(self, _origin: Self::Origin) -> DispatchResult<Self::Error> {
+            Ok(())
+        }
+    }
+
     impl identity::Trait for Test {
         type Event = ();
+        type Proposal = IdentityProposal;
     }
 
     impl asset::Trait for Test {
