+++ runtime/src/staking.rs
@@ -1817,7 +1817,11 @@ mod tests {
         Perbill,
     };
     use srml_support::traits::FindAuthor;
-    use srml_support::{assert_ok, impl_outer_origin, parameter_types};
+    use srml_support::{
+        assert_ok,
+        dispatch::{DispatchError, DispatchResult},
+        impl_outer_origin, parameter_types,
+    };
     use substrate_primitives::{Blake2Hasher, H256};
     use system::EnsureSignedBy;
 
@@ -1947,8 +1951,24 @@ mod tests {
         type Identity = identity::Module<Test>;
     }
 
+    #[derive(codec::Encode, codec::Decode, Debug, Clone, Eq, PartialEq)]
+    pub struct IdentityProposal {
+        pub dummy: u8,
+    }
+
+    impl sr_primitives::traits::Dispatchable for IdentityProposal {
+        type Origin = Origin;
+        type Trait = Test;
+        type Error = DispatchError;
+
+        fn dispatch(self, _origin: Self::Origin) -> DispatchResult<Self::Error> {
+            Ok(())
+        }
+    }
+
     impl identity::Trait for Test {
         type Event = ();
+        type Proposal = IdentityProposal;
     }
 
     parameter_types! {
