+++ runtime/src/test/mod.rs
@@ -1,0 +1,2 @@
+pub mod storage;
+pub use storage::TestStorage;
